Complete task 3 + add macOS flag to avoid getting BAC_ACCESS due to 64-bit memory architecture

+++ _xtoa.h
@@ -9,7 +9,7 @@
 
 #include "_stdtype.h"
 
-void _utoa(u32 n, char* buf, int next, u8 base);
+void _utoa(u32 n, char* buf, i8 next, u8 base);
 
 void System_itoa(i32 i, char* buf);
